% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIMSeq.R
\docType{package}
\name{PIMseq-package}
\alias{PIMseq-package}
\title{PIMSeq package}
\description{
Probablistic index models (PIM) for testing differential gene expression in RNA-seq data.
}
\author{
Alemu Takele Assefa \email{alemutak@hotmail.com}
}
