% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPIrank.R
\name{plotPIrank}
\alias{plotPIrank}
\title{PI rank plot}
\usage{
plotPIrank(pim.res, contrast, add.PI.interval = TRUE, FDR.thrld = 0.05,
  PI.thrld = c(0.4, 0.6), xlab = "gene rank",
  ylab = "probabilstic index", add.legend = TRUE, cols.seg = NULL,
  las = 1, cex.lab = 1.25, ylim = c(0, 1), col = NULL, pch = 20,
  cex = 0.5, lty.PI = 3, lwd.PI = 1.25, col.PI = "blue",
  legend.pos = "bottomright", pt.cex.legend = 1.25, cex.legend = 1,
  ...)
}
\arguments{
\item{pim.res}{outputs from \emph{PIMSeq()} function}

\item{contrast}{a vector of coefficients for the linear combination of the factors}

\item{add.PI.interval}{a logical value to add error bar}

\item{FDR.thrld}{FDR threshold, a numerical value between 0 and 1}

\item{PI.thrld}{a vector of numerical values for the down and up regulated genes PI threshold}

\item{xlab}{xlab parameter for \emph{plot()}}

\item{ylab}{ylab parameter for \emph{plot()}}

\item{add.legend}{a logical value to add a legend}

\item{cols.seg}{a vector of colors for the error bars. Size must be equal to the number of genes}

\item{las}{las parameter for \emph{plot()}}

\item{cex.lab}{cex.lab parameter for \emph{plot()}}

\item{ylim}{ylim parameter for \emph{plot()}}

\item{col}{col parameter for \emph{plot()}}

\item{pch}{pch parameter for \emph{plot()}}

\item{cex}{cex parameter for \emph{plot()}}

\item{lty.PI}{lty parameter for \emph{plot()}, for PI horizontal line(s)}

\item{lwd.PI}{lwd parameter for \emph{plot()}, for PI horizontal line(s)}

\item{col.PI}{col parameter for \emph{plot()}, for PI horizontal line(s)}

\item{legend.pos}{legend position}

\item{pt.cex.legend}{pch size for legend}

\item{cex.legend}{size of legend}

\item{...}{other graphical parameters for \emph{plot()}}
}
\value{
plot
}
\description{
Plots gene ranking based on the estimated PI. Low-regulated genes (PI <0.5) appear
in the left edge and up-regulated genes appear in the right edge. Error bars indicate the (1-alpha)*100%
confidence interval estimate of the PI.
}
\examples{
# see examples in PIMSeq() function

}
