% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testContrast.R
\name{testPIMcontrast}
\alias{testPIMcontrast}
\title{Test contrast from PIMSeq results}
\usage{
testPIMcontrast(pim.res, contrasts, only.conditions = TRUE,
  gene.warning.return = FALSE, ...)
}
\arguments{
\item{pim.res}{PIMSeq output}

\item{contrasts}{a vector of coefficients for the linear combination of the factors}

\item{only.conditions}{a logical value If only.conditions=TRUE (default), contrast will
be tested among the coefficients of condition variables only. If only.conditions=FALSE,
contrast will be tested considering all coefiecients in the model (conditions and
nuisance variables). The dimension of the contrast matrix/vector depends on this argument.}

\item{gene.warning.return}{a logical value to show warning message if the contrast 
test for a particular gene encounters an error. Default is FALSE.}

\item{...}{additional arguiments}
}
\value{
a data frame containg the estimated linear factor, test statistic, standard 
error, p value, estimated PI, and FDR adjusted p value.
}
\description{
Tests contrasts using results obtained from PIMSeq function. This function
can be called independently after fitting PIM for gene expression data.
}
\examples{
 # examples
}
